+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import type { Metadata, Viewport } from "next";
 import localFont from "next/font/local";
 
-import "./globals.css";
+import "@/app/globals.css";
 import "@/types/common";
 import "@/utils/common";
 import ReactQueryProvider from "@/reactQueryProvider";
