+++ src/domains/auth/hooks/useAuth.ts
@@ -1,6 +1,6 @@
 import { useRouter } from "next/navigation";
-import { useOAuth } from "./useOAuth";
-import { useSignIn } from "./useSignIn";
+import { useOAuth } from "./useOAuthMutation";
+import { useSignIn } from "./useSignInMutation";
 import type { OAuthProvider } from "@/domains/auth/types/entity";
 import useUserStore from "@/store/user";
