--- src/domains/jjikboul/hooks/useJjikboulDetailQuery2.ts
@@ -1,10 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import { jjikboulApi } from "@/domains/jjikboul/api/jjikboulApi";
-
-export const useGetJjikboulDetailQuery = (jjikboulId: string) => {
-  return useQuery({
-    queryKey: ["jjikboul", jjikboulId],
-    queryFn: () => jjikboulApi.getJjikboulDetail(jjikboulId),
-    enabled: !!jjikboulId,
-  });
-};
