+++ src/domains/profile/components/EditProfile.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { createContext, useContext, useState, type ReactNode } from "react";
-import { useEditProfile } from "@/domains/profile/hooks/useEditProfile";
+import { useEditProfile } from "@/domains/profile/hooks/useEditProfileMutation";
 
 interface EditProfileContextValue {
   nickname: string;
