+++ src/domains/profile/hooks/useMyProfileQuery.ts
@@ -1,13 +1,9 @@
 import { useSuspenseQuery } from "@tanstack/react-query";
 import { profileApi } from "@/domains/profile/api/profileApi";
 
-export const useMyProfile = () => {
-  const { data: profile } = useSuspenseQuery({
-    queryKey: ["profile", "my"],
+export const useMyProfileQuery = () => {
+  return useSuspenseQuery({
+    queryKey: ["profile", "me"],
     queryFn: () => profileApi.getMyProfile(),
   });
-
-  return {
-    profile,
-  };
 };
