--- src/domains/profile/hooks/useMyProfileQuery2.ts
@@ -1,9 +1,0 @@
-import { useQuery } from "@tanstack/react-query";
-import { profileApi } from "@/domains/profile/api/profileApi";
-
-export const useGetMyProfileQuery = () => {
-  return useQuery({
-    queryKey: ["profile", "me"],
-    queryFn: () => profileApi.getMyProfile(),
-  });
-};
