+++ src/domains/record/hooks/useReportReasonQuery.ts
@@ -1,8 +1,8 @@
-import { useQuery } from "@tanstack/react-query";
+import { useSuspenseQuery } from "@tanstack/react-query";
 import { recordApi } from "@/domains/record/api/recordApi";
 
 export const useReportReasonQuery = () => {
-  return useQuery({
+  return useSuspenseQuery({
     queryKey: ["report", "reason"],
     queryFn: recordApi.getReportReasons,
     staleTime: 1000 * 60 * 10, // 10분
